Extracts a GoblinHunter class.
Also moves the usage of the hit sound (hit.wav) into the Enemy (goblin)
class, instead of just hanging out in the game class.
